fixed issue with the speed value not being a double value

+++ src/main/java/pt/iceman/comsinstrumentcluster/dashboard/Dashboard.java
@@ -243,7 +243,7 @@ public abstract class Dashboard extends Screen {
                 setHighBeams((boolean) baseCommand.getValue());
                 break;
             case "speed":
-                setSpeed((double) baseCommand.getValue());
+                setSpeed((long) baseCommand.getValue());
                 break;
             case "total-distance":
                 setTotalDistance((double) baseCommand.getValue());
